requirements/base.txt
litestar[standard]
pydantic-settings
pydantic
uvicorn
tortoise-orm[asyncpg]
aerich
